+++ lib/new_relic/agent/configuration/default_source.rb
@@ -4,6 +4,7 @@
 
 require 'forwardable'
 require_relative '../../constants'
+require_relative 'sampler_config_validator'
 
 module NewRelic
   module Agent
@@ -280,16 +281,6 @@ module NewRelic
           end
           constants.compact!
           constants
-        end
-
-        def self.enforce_fallback(allowed_values: nil, fallback: nil)
-          proc do |configured_value|
-            if allowed_values.any? { |v| v =~ /#{configured_value}/i }
-              configured_value
-            else
-              fallback
-            end
-          end
         end
       end
 
@@ -1463,19 +1454,68 @@ module NewRelic
           :allowed_from_server => true,
           :description => 'Distributed tracing lets you see the path that a request takes through your distributed system. Enabling distributed tracing changes the behavior of some New Relic features, so carefully consult the [transition guide](/docs/transition-guide-distributed-tracing) before you enable this feature.'
         },
+        :'distributed_tracing.sampler.root' => {
+          :default => 'default',
+          :public => true,
+          :type => String,
+          :allowed_from_server => false,
+          :allowlist => %w[default adaptive always_on always_off trace_id_ratio_based],
+          :transform => SamplerConfigValidator.validate_sampler_strategy_with_ratio(
+            :'distributed_tracing.sampler.root',
+            :'distributed_tracing.sampler.root.trace_id_ratio_based.ratio'
+          ),
+          :description => 'This setting controls the behavior of transaction sampling for transactions without a remote parent, traces that originate within this instance of the New Relic agent. Available values are `default`, `adaptive`, `always_on`, `always_off`, and `trace_id_ratio_based`. At this time `default` and `adaptive` are the same.'
+        },
         :'distributed_tracing.sampler.remote_parent_sampled' => {
           :default => 'default',
           :public => true,
           :type => String,
-          :allowed_from_server => true,
-          :description => 'This setting controls the behavior of transaction sampling when a remote parent is sampled and the trace flag is set in the traceparent. Available values are `default`, `always_on`, and `always_off`.'
+          :allowed_from_server => false,
+          :allowlist => %w[default adaptive always_on always_off trace_id_ratio_based],
+          :transform => SamplerConfigValidator.validate_sampler_strategy_with_ratio(
+            :'distributed_tracing.sampler.remote_parent_sampled',
+            :'distributed_tracing.sampler.remote_parent_sampled.trace_id_ratio_based.ratio'
+          ),
+          :description => 'This setting controls the behavior of transaction sampling when a remote parent is sampled. Available values are `default`, `adaptive`, `always_on`, `always_off`, and `trace_id_ratio_based`. At this time `default` and `adaptive` are the same.'
         },
         :'distributed_tracing.sampler.remote_parent_not_sampled' => {
           :default => 'default',
           :public => true,
           :type => String,
-          :allowed_from_server => true,
-          :description => 'This setting controls the behavior of transaction sampling when a remote parent is not sampled and the trace flag is not set in the traceparent. Available values are `default`, `always_on`, and `always_off`.'
+          :allowed_from_server => false,
+          :allowlist => %w[default adaptive always_on always_off trace_id_ratio_based],
+          :transform => SamplerConfigValidator.validate_sampler_strategy_with_ratio(
+            :'distributed_tracing.sampler.remote_parent_not_sampled',
+            :'distributed_tracing.sampler.remote_parent_not_sampled.trace_id_ratio_based.ratio'
+          ),
+          :description => 'This setting controls the behavior of transaction sampling when a remote parent is not sampled. Available values are `default`, `adaptive`, `always_on`, `always_off`, and `trace_id_ratio_based`. At this time `default` and `adaptive` are the same.'
+        },
+        :'distributed_tracing.sampler.root.trace_id_ratio_based.ratio' => {
+          :default => nil,
+          :public => true,
+          :type => Float,
+          :allow_nil => true,
+          :allowed_from_server => false,
+          :transform => SamplerConfigValidator.method(:validate_sampling_ratio),
+          :description => 'The ratio used for the trace_id_ratio_based sampling decision for the root sampler. This must be a float between 0.0 and 1.0. If you provide an invalid value, the sampler will not use the trace_id_ratio_based sampler and will return to the default behavior. If you do not provide a value, the sampler will not use the trace_id_ratio_based_sampler and fall back to the default sampler.'
+        },
+        :'distributed_tracing.sampler.remote_parent_sampled.trace_id_ratio_based.ratio' => {
+          :default => nil,
+          :public => true,
+          :type => Float,
+          :allow_nil => true,
+          :allowed_from_server => false,
+          :transform => SamplerConfigValidator.method(:validate_sampling_ratio),
+          :description => 'The ratio used for the trace_id_ratio_based sampling decision for the remote parent sampled sampler. This must be a float between 0.0 and 1.0. If you provide an invalid value, the sampler will not use the trace_id_ratio_based sampler and will return to the default behavior. If you do not provide a value, the sampler will not use the trace_id_ratio_based_sampler and fall back to the default sampler.'
+        },
+        :'distributed_tracing.sampler.remote_parent_not_sampled.trace_id_ratio_based.ratio' => {
+          :default => nil,
+          :public => true,
+          :type => Float,
+          :allow_nil => true,
+          :allowed_from_server => false,
+          :transform => SamplerConfigValidator.method(:validate_sampling_ratio),
+          :description => 'The ratio used for the trace_id_ratio_based sampling decision for the remote parent not sampled sampler. This must be a float between 0.0 and 1.0. If you provide an invalid value or do not provide a value, the sampler will not use the trace_id_ratio_based_sampler and fall back to the default sampler.'
         },
         # Elasticsearch
         :'elasticsearch.capture_cluster_name' => {